backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-multipart==0.0.20
pydantic==2.11.7
pydantic-ai==0.3.5
PyPDF2==3.0.1
aiofiles==24.1.0
sentence-transformers==4.1.0
numpy==2.0.2
openai==1.91.0
python-dotenv==1.1.1

# RAG-Anything Integration - DO NOT modify existing entries above
raganything>=0.1.0
lightrag-hku>=0.1.0
neo4j>=5.0.0